migration fix2

+++ db/migrate/20230509052843_drop_guests_table.rb
@@ -1,5 +1,5 @@
 class DropGuestsTable < ActiveRecord::Migration[7.0]
   def change
-    drop_table :guests
+    drop_table :guests if table_exists?(:guests)
   end
 end
